UBUNTU: SAUCE: intel_pstate -- toggle default to disable
The intel_pstate driver is most unhappy on 3.9 and 3.10, for now disable
it, but do so in such a way we can turn it on for testing on the kernel
command line.

It seems likely we are also expected to have a userspace component for
fan control is we are using pstate_idle.  This seems suboptimal:

    https://01.org/linux-thermal-daemon

BugLink: http://bugs.launchpad.net/bugs/1188647
Signed-off-by: Andy Whitcroft <apw@canonical.com>