UBUNTU: [Packaging] tools -- hv tools build correctly against the built headers
We are forced to hack the header paths in the source to build the
hv_*_daemons, and in doing so also trigger this to be built against
the headers installed on the system rather than the ones we are
building right now which may differ.  To avoid this epic fail depend on
install-arch-headers which is making the headers anyhow, and use the ones
it makes.  Also directly build the binaries we care about avoiding the
need to add a Makefile.

Signed-off-by: Andy Whitcroft <apw@canonical.com>