UBUNTU: SAUCE: (no-up) apparmor: fix refcount bug in apparmor pivotroot
This is a fix to code that is not upstream.

When the pivotroot check was refactored into the callback fn, putting
of the target fn was not pushed down because target was referenced in
by audit_log in the error path. However target will never be set in
the cases when audit_log is called from the error path.

So instead of passing target the target value back out of profile_pivotroot
push putting the target reference down as it is not needed in aa_pivotroot.

BugLink: http://bugs.launchpad.net/bugs/1308765

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>