UBUNTU: [Config]: Add missing modules to inclusion list
Compared to Trusty some drivers were missing which turned out to
be paravirt drivers that really are needed in the cloud-images
which do not install the extra package.
I also added all of drivers/vhost because that whole directory
sounds like drivers we may get asked for (one was included before).
Unfortunately that requires target core which on a second thought
might be something that actually fits into the virt host scheme.

Signed-off-by: Stefan Bader <smb@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>