tcm_loop: Show address of tpg in configfs
In the past, the scsi_host's number wasn't shown anywhere, user have to
depends on vpg_83 to find the block device, which is also depends on
backstore implementation.

It's better for tcm_loop to provide necessary straightforward information
on locate the block device it created.

This patch would help to locate the block device created by tcm_loop.  The
address would be shown at e.g.

/sys/kernel/config/target/loopback/naa.60014059436855c1/tpgt_1/address

which would looks like "2:0:1", and the lun number can be found at
"/tpgt_1/lun/lun_0".  Altogether they formated the scsi address of device
as "2:0:1:0", which can be used to locate the device easily through
'lsscsi'.

Signed-off-by: Sheng Yang <sheng@yasker.org>